Move query model introspection functions to new module
