+++ ehrql/query_engines/csv.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from ehrql.query_engines.in_memory import InMemoryQueryEngine
 from ehrql.query_engines.in_memory_database import InMemoryDatabase
-from ehrql.query_model.nodes import get_table_nodes
+from ehrql.query_model.introspection import get_table_nodes
 from ehrql.utils.orm_utils import (
     orm_classes_from_tables,
     read_orm_models_from_csv_directory,
